device/asus/tf101: Add support for low fps boot animation. Some devices can not support over 100+ fps, leaving too much black screen time during boot.
To use, add flag 'USE_LOWFPS_BOOTANI := true' to your omni_<device>.mk file, ahead of the call to vendor/omni/config/common.mk (or common_tablet.mk). 

This was tested on clean builds for tf101 and with flag (and corresponding vendor patch) the low fps boot animation was pulled in and without the flag the normal boot animation was pulled in. Should work for all devices.

Change-Id: I219df9a01b4c9f494879d584a3433213464617ea